cherry-pick 01builders/nova@376e9cfe5d84b359b05b60f118b8f36dc33a0f62
