+++ multiplexer/cmd/passthrough_test.go
@@ -1,4 +1,4 @@
-package multiplexer
+package cmd
 
 import (
 	"bytes"
